fix: track point heart rate 자료형 변경, Elevation gain nullabl처리

+++ lib/features/workout_history/domain/entities/track_point.dart
@@ -13,7 +13,7 @@ class TrackPoint {
   final double latitude;
   final double longitude;
   final double? speed; // km/h
-  final int? heartRate; // bpm
+  final double? heartRate; // bpm
 
   @override
   bool operator ==(Object other) {
